Bolt: Optimize NumPy array creation and scaling operations
- Use `np.asarray()` instead of `np.array()` in `tensor_from_pil` to avoid unnecessary deep copying of PIL images.
- Explicitly use `np.float64()` constants when scaling `np.float64` arrays in `processing_context.py` to prevent dynamic implicit type resolution overhead during operations.

+++ src/nodetool/workflows/processing_context.py
@@ -1460,7 +1460,7 @@ class ProcessingContext:
         if dtype == "float32" and samples.dtype == np.int16:
             samples = samples.astype(np.float32) / np.float32(32768.0)
         elif dtype == "float64" and samples.dtype == np.int16:
-            samples = samples.astype(np.float64) / 32768.0
+            samples = samples.astype(np.float64) / np.float64(32768.0)
         elif dtype == "int16" and samples.dtype in (np.float32, np.float64):
             samples = (samples * 32768.0).clip(-32768, 32767).astype(np.int16)
         elif dtype != str(samples.dtype):
